+++ routes/actions.php
@@ -1,6 +1,6 @@
 <?php
 
-use Aerni\Zipper\ZipperController;
+use Aerni\Zipper\Http\Controllers\ZipperController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/create/{id}', [ZipperController::class, 'create'])->name('zipper.create');
